+++ src/main/java/org/thanhlong/Midterm/Repository/RoleRepository.java
@@ -5,6 +5,5 @@ import org.springframework.stereotype.Repository;
 import org.thanhlong.Midterm.Models.Role;
 
 @Repository
-public interface RoleRepository extends JpaRepository<Role, Integer> {
-
+public interface RoleRepository extends JpaRepository<Role ,Long> {
 }
