+++ src/main/java/org/thanhlong/Midterm/Service/RoleService.java
@@ -1,6 +1,5 @@
 package org.thanhlong.Midterm.Service;
 
-
 import org.springframework.stereotype.Service;
 import org.thanhlong.Midterm.Models.Role;
 
@@ -10,6 +9,5 @@ import java.util.Optional;
 @Service
 public interface RoleService {
     List<Role> getAllRole();
-    Optional<Role> findRoleById(int id);
-
+    Optional<Role> findRoleById(long id);
 }
